[Update] Move main to game_ui
